Generate-images-using-Amazon-Bedrock-with-stability-diffusion-model/requirements.txt
streamlit==1.25.0
boto3==1.28.0
Pillow==9.2.0